requirements/requirements.txt
flask
flask-restx
Flask-PyMongo
dnspython
python-dotenv